// transit.mobile/models/login-response-model.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export interface LoginResponseModel {
    token: string;
    refreshToken: string;
    expires: Date;
    username: string;
    email: string;
    firstName: string;
    lastName: string;
}